# Remove parallel acquirement of locks for `MultiFrame` witness caching