@arthurpaulino 2023-12-07
!gpu-benchmark