@arthurpaulino 2023-12-11
1. It's not that the caching loop is slower with the locks. The problem is that the locks from caching can cause a slowdown on the folding steps
2. If we merge the two, we prevent folding from happening in parallel as the witnesses are generated.

To picture this, imagine a computation with 1000 multiframes. The way current main works is by locking the last 999 steps for parallel witness generation and let folding start right away. And the issue becomes clearer here because the probability of the witness for the second step becoming ready in time is pretty low, not only because there are 999 other multiframes to have their witnesses generated but also because the resources are more sparse.

This PR makes witness generation be focused so folding has a lower chance of having to wait. And this is key to achieve a better utilization of the GPU.