@arthurpaulino 2023-12-11
> I think what you mean is that the thread consumption of parallel caching is exhausting resources for folding.

While that's true, it's not what I meant. I meant that spending resources on the witness generation for the 1000th multiframe (and on all 997 multiframes before that one) won't help with making the second proving step move forward. In fact, the more multiframes we have, the higher are the chances of the second proving step having to wait. And making GPU wait is pretty bad.