@arthurpaulino 2023-12-11
Indeed, I am not touching the looks. I am making earlier locks be released sooner.